Construct kernels of LALR(1) collection of sets of items
